feat: update deps and code to v2 and eslint

+++ functions/src/firestore/request/deleteClubRequest.ts
@@ -1,19 +1,21 @@
-/* eslint-disable linebreak-style */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-/* eslint-disable require-jsdoc */
+
 /* eslint-disable max-len */
-import * as functions from "firebase-functions";
-import firebaseDAO from "../../firebaseSingleton";
-import {QueryDocumentSnapshot} from "firebase-functions/lib/providers/firestore";
+
+import firebaseDAO from '../../firebaseSingleton';
+import {FirestoreEvent, QueryDocumentSnapshot} from 'firebase-functions/v2/firestore';
 
 const db = firebaseDAO.instance.db;
 // const auth = firebaseDAO.instance.auth;
 
-export async function deleteClubRequest(snapshot: QueryDocumentSnapshot, context: functions.EventContext) {
-  console.log("deleteClubRequest");
-  const userId = context.params.userId;
-  const clubId = context.params.clubId;
+export async function deleteClubRequest(event: FirestoreEvent<QueryDocumentSnapshot | undefined>) {
+  console.log('deleteClubRequest');
+  const userId = event.params.userId;
+  const clubId = event.params.clubId;
 
-  return db.collection("club").doc(clubId).collection("requests").doc(userId).delete();
+  if (!event.data) {
+    console.log('No data associated with the Club Request');
+    return;
+  }
+
+  return db.collection('club').doc(clubId).collection('requests').doc(userId).delete();
 }
