+++ functions/src/firestore/request/deleteTeamRequest.ts
@@ -1,19 +1,21 @@
-/* eslint-disable linebreak-style */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-/* eslint-disable require-jsdoc */
+
 /* eslint-disable max-len */
-import * as functions from "firebase-functions";
-import firebaseDAO from "../../firebaseSingleton";
-import {QueryDocumentSnapshot} from "firebase-functions/lib/providers/firestore";
+
+import firebaseDAO from '../../firebaseSingleton';
+import {FirestoreEvent, QueryDocumentSnapshot} from 'firebase-functions/v2/firestore';
 
 const db = firebaseDAO.instance.db;
 // const auth = firebaseDAO.instance.auth;
 
-export async function deleteTeamRequest(snapshot: QueryDocumentSnapshot, context: functions.EventContext) {
-  console.log("deleteTeamRequest");
-  const userId = context.params.userId;
-  const teamId = context.params.teamId;
+export async function deleteTeamRequest(event: FirestoreEvent<QueryDocumentSnapshot | undefined>) {
+  console.log('deleteTeamRequest');
+  const userId = event.params.userId;
+  const teamId = event.params.teamId;
 
-  return db.collection("teams").doc(teamId).collection("requests").doc(userId).delete();
+  if (!event.data) {
+    console.log('No data associated with the Team Request');
+    return;
+  }
+
+  return db.collection('teams').doc(teamId).collection('requests').doc(userId).delete();
 }
